+++ docs/misc/changelog.md
@@ -2,11 +2,11 @@
 
 # Changelog
 
-## Release 2.9.0a1 (WIP)
+## Release 2.9.0a2 (WIP)
 
 ### Breaking Changes:
 - Relaxed Gymnasium version range (from `"gymnasium>=0.29.1,<1.3.0"` to `"gymnasium>=0.29.1,<2.0"`)
-- `pandas` and `matplotlib` are no longer core dependencies; they are now optional and only only required for plotting functionality (moved to `stable_baselines3[extra]`).
+- `pandas` and `matplotlib` are no longer core dependencies; they are now optional and only required for loading results and plotting (moved to `stable-baselines3[extra]`).
 - Moved `read_json` and `read_csv` helper functions to test files
 
 ### New Features: